Moving firebase.ts and authHelper.ts to utils
