+++ frontend/src/utils/authHelper.ts
